Initial implementation of baseurl parameter for zip endpoint.
